feat: add aslint to task
