+++ .eslintrc.json
@@ -1,0 +1,19 @@
+{
+   "env": {
+      "browser": true,
+      "commonjs": true,
+      "es2021": true
+   },
+   "extends": "airbnb-base",
+   "parserOptions": {
+      "ecmaVersion": "latest",
+      "sourceType": "module"
+   },
+   "rules": {
+      "indent": 0,
+      "import/extensions": [
+         "error",
+         "always"
+      ]
+   }
+}
