+++ index.html
@@ -10,7 +10,7 @@
 </head>
 
 <body>
-   <script src="./main.js"></script>
+   <script type="module" src="./main.js"></script>
 </body>
 
 </html>
