+++ modules/variables.js
@@ -1,0 +1,7 @@
+export const unactive = (item) => {
+   item.classList.remove('active');
+};
+
+export const active = (item) => {
+   item.classList.add('active');
+};
